+++ CHANGES.rst
@@ -34,6 +34,9 @@ Changelog
   [gbastien]
 - Fixed `utils.sendMailIfRelevant` when `isPermission=True` that was simply broken.
   [gbastien]
+- Changed behavior of `MeetingItem.get_representatives_in_charge`, it will return
+  `held_position objects`, no more the `MeetingItem.groupsInCharge organizations`.
+  [gbastien]
 
 4.2rc21 (2022-03-22)
 --------------------